# pkg/api/access.go
/*
 *    Copyright 2020 InfAI (CC SES)
 *
 *    Licensed under the Apache License, Version 2.0 (the "License");
 *    you may not use this file except in compliance with the License.
 *    You may obtain a copy of the License at
 *
 *        http://www.apache.org/licenses/LICENSE-2.0
 *
 *    Unless required by applicable law or agreed to in writing, software
 *    distributed under the License is distributed on an "AS IS" BASIS,
 *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 *    See the License for the specific language governing permissions and
 *    limitations under the License.
 */

package api

import (
	"errors"

	"github.com/SENERGY-Platform/authorization/pkg/api/util"
	"github.com/SENERGY-Platform/authorization/pkg/authorization"
	"github.com/SENERGY-Platform/authorization/pkg/configuration"
	"github.com/SENERGY-Platform/authorization/pkg/model"
	"github.com/gin-gonic/gin"
	"github.com/ory/ladon"
)

func init() {
	endpoints = append(endpoints, AccessEndpoint)
}

type KongMessage struct {
	Result bool   `json:"result"`
	Error  string `json:"error"`
}

func AccessEndpoint(router *gin.Engine, _ configuration.Config, _ util.Jwt, guard *authorization.Guard) {
	router.GET("/access", accessHandler(guard))
}

// accessHandler godoc
// @Summary Check access via Ladon policy
// @Description Evaluates a Ladon policy request and returns whether access is allowed.
// @Tags access
// @Accept json
// @Produce json
// @Param request body ladon.Request true "Ladon access request"
// @Success 200 {object} KongMessage
// @Failure 400 {string} ErrorResponse
// @Router /access [get]
func accessHandler(guard *authorization.Guard) gin.HandlerFunc {
	return func(c *gin.Context) {
		var ladonRequest ladon.Request

		err := c.ShouldBindJSON(&ladonRequest)
		if err != nil {
			c.Error(errors.Join(model.ErrBadRequest, err))
			return
		}

		var message KongMessage
		authErr := guard.IsAllowed(&ladonRequest)
		if authErr != nil {
			message = KongMessage{
				false,
				authErr.Error(),
			}
		} else {
			message = KongMessage{
				true,
				"",
			}
		}

		c.JSON(model.GetStatusCode(nil), message)
	}
}